AppMessage classic PebbleKit Android implementation (#16)
* generalize PKJSLifecycleManager into CompanionAppLifecycleManager

* add AppMessage PebbleKit classic support

* Fix inboundAppMessages not emitting early messages

when consumeAsFlow is called, it immediately hogs all the messages, which means that the Buffer in the Channel above has no effect - if nobody is collecting the flow, the messages are lost. This is fixed by moving away from SharedFlow to regular flow and by using receive instead of consume.

A side effect of this is that this flow cannot be collected by more than one consumer at once, but from what I can see, there is no use case for that?

* Emit AppMessage unsigned numbers as ULong

This allows PebbleKit Android to distinguish between incoming signed and unsigned values. It should not affect JS, because it is a normal number behind the scenes

* Add PebbleKit start/stop listeners

* Add PebbleKit Classic provider

* Use ordered intents for reduced latency

* Fix PebbleKitProvider throwing errors on startup

LibPebble might not be initialised yet at the point when the first requests arrive, so we now handle this case more gracefully. Updating has been moved to its own class PebbleKitProviderNotifier, which is created after the LibPebble fully inits

* Cache AppMessages per-app

this fixes the following scenario:

1. App A is active, with PBJS active
2. App B is activated and immediately sends an AppMessage
3. Afterward, AppRunStateStart is sent for the app switch
4. Packet from the step 2 is dropped, because PBJS for the app A is still active

* handle app not declaring PebbleKit provider

* add BroadcastReceiver exported flag, set for PebbleKit

* use old getSerializableExtra to support SDK < 33

* Clarify PebbleKit iOS status

* Create PebbleKitClassic classes with constructor injection

* Update coroutineScope naming

* Fix iOS build

* Fix tests

---------

+++ libpebble3/src/commonMain/kotlin/io/rebble/libpebblecommon/packets/AppMessage.kt
@@ -75,7 +75,7 @@ class AppMessageTuple() : StructMappable() {
             }.valueNumber
         }
 
-    val dataAsUnsignedNumber: Long
+    val dataAsUnsignedNumber: ULong
         get() {
             val obj = when (val size = dataLength.get().toInt()) {
                 1 -> SUByte(StructMapper())
@@ -85,7 +85,7 @@ class AppMessageTuple() : StructMappable() {
             }
             return obj.apply {
                 fromBytes(DataBuffer(data.get()))
-            }.valueNumber
+            }.valueNumber.toULong()
         }
 
     override fun toString(): String {
